Fix java-objc-natives causing an import to fail
This commit adds a rather hacky fix, if you tried to use
PufferfishGradle before this patch, gradle would not be able to resolve
java-objc-bridge. This just adds a check if the library name contains
java-objc-bridge, if it does, the library will not be added to the
dependency list.

This should probably be fixed another way... but it's fine for now :)